refactor(sql-playground): move models to feature folder
- Move 11 TypeScript interfaces to features/sql-playground/models/
  (Constraint, Database, QueryTab, Routine, SQLExecuteResponse,
   SQLResponse, Table, Template, TemplateCategory, Trigger, View)
- Update 53 import statements across 21 files:
  * service/sql-playground.service.ts
  * page-components/sql-playground/** (components + state)
  * dialogs/new-sql-template/
- Delete old model/sql_playground/ directory
- No functional changes

Step 3 of 6
