+++ modules/fbs-core/web/src/app/features/sql-playground/models/QueryTab.ts
@@ -1,6 +1,6 @@
-import { Course } from "../Course";
-import { Task } from "../Task";
-import { BackendUser } from "../../page-components/sql-playground/collab/backend.service";
+import { Course } from "../../../model/Course";
+import { Task } from "../../../model/Task";
+import { BackendUser } from "../../../page-components/sql-playground/collab/backend.service";
 
 export interface QueryTab {
   id: string;
