+++ modules/fbs-core/web/src/app/features/sql-playground/models/Template.ts
