+++ modules/fbs-core/web/src/app/page-components/sql-playground/db-control-panel/db-control-co-working/db-control-co-working.component.ts
@@ -1,7 +1,7 @@
 import { Component, OnInit } from "@angular/core";
 import { Store } from "@ngrx/store";
 import { Observable } from "rxjs";
-import { Database } from "../../../../model/sql_playground/Database";
+import { Database } from "src/app/features/sql-playground/models/Database";
 import { MatSnackBar } from "@angular/material/snack-bar";
 import { AuthService } from "src/app/service/auth.service";
 import { MatDialog } from "@angular/material/dialog";
