+++ modules/fbs-core/web/src/app/page-components/sql-playground/sql-playground.component.ts
@@ -1,11 +1,11 @@
 import { AfterViewChecked, Component, OnInit } from "@angular/core";
 import { Store } from "@ngrx/store";
 import { Subject, BehaviorSubject, Observable } from "rxjs";
-import { Routine } from "src/app/model/sql_playground/Routine";
-import { Trigger } from "src/app/model/sql_playground/Trigger";
-import { View } from "src/app/model/sql_playground/View";
-import { Table } from "src/app/model/sql_playground/Table";
-import { Constraint } from "src/app/model/sql_playground/Constraint";
+import { Routine } from "src/app/features/sql-playground/models/Routine";
+import { Trigger } from "src/app/features/sql-playground/models/Trigger";
+import { View } from "src/app/features/sql-playground/models/View";
+import { Table } from "src/app/features/sql-playground/models/Table";
+import { Constraint } from "src/app/features/sql-playground/models/Constraint";
 import { TitlebarService } from "../../service/titlebar.service";
 import { AuthService } from "src/app/service/auth.service";
 import { MatSnackBar } from "@angular/material/snack-bar";
