+++ modules/fbs-core/web/src/app/page-components/sql-playground/state/sql-playground.actions.ts
@@ -1,9 +1,9 @@
 import { createAction, props } from "@ngrx/store";
-import { Routine } from "src/app/model/sql_playground/Routine";
-import { Trigger } from "src/app/model/sql_playground/Trigger";
-import { View } from "src/app/model/sql_playground/View";
-import { Table } from "src/app/model/sql_playground/Table";
-import { Constraint } from "src/app/model/sql_playground/Constraint";
+import { Routine } from "src/app/features/sql-playground/models/Routine";
+import { Trigger } from "src/app/features/sql-playground/models/Trigger";
+import { View } from "src/app/features/sql-playground/models/View";
+import { Table } from "src/app/features/sql-playground/models/Table";
+import { Constraint } from "src/app/features/sql-playground/models/Constraint";
 import {
   BackendDefintion,
   DatabaseInformation,
